pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "voicepipe"
version = "0.1.0"
description = "Voice recording and transcription daemon with systray integration"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/yourusername/voicepipe"
repository = "https://github.com/yourusername/voicepipe"
keywords = ["voice", "recording", "transcription", "cli", "audio"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
packages = [{include = "voicepipe"}]
include = ["voicepipe/assets/*"]

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.0"
openai = "^1.0"
python-dotenv = "^1.0"
sounddevice = "^0.4.6"
numpy = "^1.24.0"
# Optional systray dependencies
pystray = {version = "^0.19.0", optional = true}
pillow = {version = "^10.0", optional = true}

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
black = "^22.0"
flake8 = "^5.0"
mypy = "^1.0"

[tool.poetry.extras]
systray = ["pystray", "pillow"]
all = ["pystray", "pillow"]

[tool.poetry.scripts]
voicepipe = "voicepipe.cli:main"
voicepipe-fast = "voicepipe.fast:main"
voicepipe-toggle = "voicepipe.fast:toggle_main"
voicepipe-transcribe-file = "voicepipe.entrypoints:transcribe_file_main"
voicepipe-transcriber-daemon = "voicepipe.transcriber_daemon:main"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.pytest.ini_options]
pythonpath = ["."]
markers = [
    "live: opt-in integration tests requiring mic/network/API keys (enable via VOICEPIPE_LIVE_TESTS=1)",
    "desktop: interactive desktop required (typing/focus tests)",
]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = ["sounddevice", "pystray", "PIL", "PIL.*"]
ignore_missing_imports = true
